add: watermark support for photos

+++ src/models/Album.ts
@@ -2,6 +2,7 @@ import connection from '../connectors/sql.connector';
 import { AlbumInterface } from '../@types/interfaces/AlbumInterface';
 import { User } from './User';
 import { ApiError } from '../errors/api.error';
+import { PhotoId } from '../@types/PhotoId';
 
 export class Album implements AlbumInterface {
     albumId?: number;
@@ -10,7 +11,7 @@ export class Album implements AlbumInterface {
     location: string;
     date: string;
     path: string;
-    photos?: string[];
+    photos?: { url: string; watermark: boolean }[];
 
     constructor(albumName: string, user: User, location: string, date: string) {
         this.albumName = albumName;
@@ -22,18 +23,9 @@ export class Album implements AlbumInterface {
 
     async save(): Promise<string> {
         try {
-            await connection.query(
-                'INSERT INTO albums (albumName, userId, location, date, path) VALUES (?);',
-                [
-                    [
-                        this.albumName,
-                        this.userId,
-                        this.location,
-                        this.date,
-                        this.path
-                    ]
-                ]
-            );
+            await connection.query('INSERT INTO albums (albumName, userId, location, date, path) VALUES (?);', [
+                [this.albumName, this.userId, this.location, this.date, this.path]
+            ]);
         } catch (err) {
             throw new ApiError(400, `Album ${this.albumName} already exists`);
         }
@@ -41,10 +33,7 @@ export class Album implements AlbumInterface {
         return `Added ${this.albumName}`;
     }
 
-    static async getAlbumId(
-        albumName: string,
-        userId: number
-    ): Promise<number> {
+    static async getAlbumId(albumName: string, userId: number): Promise<number> {
         try {
             const albumData = await this.getAlbumData(albumName, userId);
             return albumData.albumId!;
@@ -53,30 +42,28 @@ export class Album implements AlbumInterface {
         }
     }
 
-    static async getAlbumData(
-        albumName: string,
-        userId: number
-    ): Promise<Album> {
-        const query = await connection.query(
-            `SELECT * FROM albums WHERE albumName=? and userId=?`,
-            [[albumName], [userId]]
-        );
+    static async getAlbumData(albumName: string, userId: number): Promise<Album> {
+        const query = await connection.query(`SELECT * FROM albums WHERE albumName=? and userId=?`, [
+            [albumName],
+            [userId]
+        ]);
         return JSON.parse(JSON.stringify(query))[0][0];
     }
 
-    static async delete(albumName: string, userId: number): Promise<string> {
-        await connection.query(
-            'DELETE FROM albums WHERE albumName=? AND userId=?',
-            [[albumName], [userId]]
-        );
-        return `Deleted ${albumName}`;
+    static async getAlbumPhotos(albumId: number): Promise<PhotoId[]> {
+        const query = (await connection.query('SELECT photoId, waterMark FROM photos WHERE albumId=?', [
+            [albumId]
+        ])) as any[];
+        return query[0];
     }
 
-    static async getUserAlbums(userId: number): Promise<Album[]> {
-        const query = (await connection.query(
-            'SELECT albumName, location, date FROM albums WHERE userId=?',
-            [[userId]]
-        )) as any[];
-        return query[0];
+    static async countPhotos(album: Album): Promise<number> {
+        const photos = await Album.getAlbumPhotos(album.albumId!);
+        return photos.length;
+    }
+
+    static async delete(albumName: string, userId: number): Promise<string> {
+        await connection.query('DELETE FROM albums WHERE albumName=? AND userId=?', [[albumName], [userId]]);
+        return `Deleted ${albumName}`;
     }
 }
